pyproject.toml
[project]
name = "edumedia-explorer"
version = "0.1.0"
description = "EduMedia Author — hierarchical subject pages, rich text, and interactive annotations"
readme = "README.md"
requires-python = ">=3.12,<3.16"
dependencies = [
    "fastapi[standard]==0.135.3",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "uvicorn==0.44.0",
    "sqlalchemy>=2.0.40",
    "beautifulsoup4>=4.13.3",
    "nh3>=0.2.21",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest==9.0.3",
    "ruff==0.15.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
